Moves Document into component
